- Moved PokemonMarkerExtended to map package.
